+++ packages/remix-react/components.tsx
@@ -638,6 +638,7 @@ function V2Meta() {
   let location = useLocation();
 
   let meta: V2_HtmlMetaDescriptor[] = [];
+  let leafMeta: V2_HtmlMetaDescriptor[] | null = null;
   let parentsData: { [routeId: string]: AppData } = {};
 
   let matchesWithMeta: RouteMatchWithMeta[] = matches.map((match) => ({
@@ -668,6 +669,11 @@ function V2Meta() {
               matches: matchesWithMeta,
             })
           : routeModule.meta;
+    } else if (leafMeta) {
+      // We only assign the route's meta to the nearest leaf if there is no meta
+      // export in the route. The meta function may return a falsey value which
+      // is effectively the same as an empty array.
+      routeMeta = leafMeta;
     }
 
     routeMeta = routeMeta || [];
@@ -686,6 +692,7 @@ function V2Meta() {
     matchesWithMeta[index].meta = routeMeta;
     meta = routeMeta;
     parentsData[routeId] = data;
+    leafMeta = meta;
   }
 
   return (
@@ -1182,6 +1189,10 @@ function convertRouterFetcherToRemixFetcher(
       let fetcher: FetcherStates["SubmittingAction"] = {
         state,
         type: "actionSubmission",
+        formMethod: formMethod.toUpperCase() as ActionSubmission["method"],
+        formAction: formAction,
+        formEncType: formEncType,
+        formData: formData,
         submission: {
           method: formMethod.toUpperCase() as ActionSubmission["method"],
           action: formAction,
@@ -1209,6 +1220,10 @@ function convertRouterFetcherToRemixFetcher(
           let fetcher: FetcherStates["ReloadingAction"] = {
             state,
             type: "actionReload",
+            formMethod: formMethod.toUpperCase() as ActionSubmission["method"],
+            formAction: formAction,
+            formEncType: formEncType,
+            formData: formData,
             submission: {
               method: formMethod.toUpperCase() as ActionSubmission["method"],
               action: formAction,
@@ -1223,13 +1238,16 @@ function convertRouterFetcherToRemixFetcher(
           let fetcher: FetcherStates["LoadingActionRedirect"] = {
             state,
             type: "actionRedirect",
+            formMethod: formMethod.toUpperCase() as ActionSubmission["method"],
+            formAction: formAction,
+            formEncType: formEncType,
+            formData: formData,
             submission: {
               method: formMethod.toUpperCase() as ActionSubmission["method"],
               action: formAction,
               encType: formEncType,
               formData: formData,
-              // TODO???
-              key: "todo-what-is-this?",
+              key: "",
             },
             data: undefined,
           };
@@ -1254,6 +1272,10 @@ function convertRouterFetcherToRemixFetcher(
         let fetcher: FetcherStates["SubmittingLoader"] = {
           state: "submitting",
           type: "loaderSubmission",
+          formMethod: formMethod.toUpperCase() as LoaderSubmission["method"],
+          formAction: formAction,
+          formEncType: formEncType,
+          formData: formData,
           submission: {
             method: formMethod.toUpperCase() as LoaderSubmission["method"],
             action: url.pathname + url.search,