+++ packages/remix-react/entry.ts
@@ -18,6 +18,9 @@ export interface EntryContext extends RemixContextObject {
 }
 
 export interface FutureConfig {
+  unstable_cssModules: boolean;
+  unstable_cssSideEffectImports: boolean;
+  unstable_vanillaExtract: boolean;
   v2_errorBoundary: boolean;
   v2_meta: boolean;
 }