+++ packages/remix-react/index.tsx
@@ -6,6 +6,7 @@ export type {
   NavigateFunction,
   Params,
   Path,
+  ShouldRevalidateFunction,
   SubmitFunction,
   SubmitOptions,
 } from "react-router-dom";
@@ -26,6 +27,7 @@ export {
   useResolvedPath,
   useRevalidator,
   useRouteError,
+  useRouteLoaderData,
   useSearchParams,
   useSubmit,
 } from "react-router-dom";
@@ -42,7 +44,6 @@ export {
   Scripts,
   Link,
   NavLink,
-  RemixEntry,
   PrefetchPageLinks,
   LiveReload,
   useTransition,
@@ -51,6 +52,7 @@ export {
   useLoaderData,
   useMatches,
   useActionData,
+  RemixContext as UNSAFE_RemixContext,
 } from "./components";
 
 export type { FormMethod, FormEncType } from "./data";
@@ -61,8 +63,8 @@ export { useCatch } from "./errorBoundaries";
 export type { HtmlLinkDescriptor } from "./links";
 export type {
   HtmlMetaDescriptor,
-  CatchBoundaryComponent,
-  RouteModules,
+  RouteModules as UNSAFE_RouteModules,
+  ShouldReloadFunction,
 } from "./routeModules";
 
 export { ScrollRestoration } from "./scroll-restoration";
@@ -72,6 +74,13 @@ export { RemixServer } from "./server";
 
 export type { Fetcher } from "./transition";
 
-export type { AssetsManifest, EntryContext } from "./entry";
-export type { RouteData } from "./routeData";
-export type { EntryRoute, RouteManifest } from "./routes";
+export type {
+  FutureConfig as UNSAFE_FutureConfig,
+  AssetsManifest as UNSAFE_AssetsManifest,
+  RemixContextObject as UNSAFE_RemixContextObject,
+} from "./entry";
+
+export type {
+  EntryRoute as UNSAFE_EntryRoute,
+  RouteManifest as UNSAFE_RouteManifest,
+} from "./routes";